+++ src/Connection/ServiceConnection.ts
@@ -70,6 +70,14 @@ export class ServiceConnection {
         this.handshakeRequested = false;
         this.handshakeCompleted = false;
 
+        this.webSocket.addEventListener('error', (ev) => console.error(ev));
+        this.webSocket.addEventListener('close', (_ev) => {
+            if (this.handshakeCompleted) {
+                TouchFree.DispatchEvent('OnServiceStatusChange', 'Disconnected');
+                console.log('Disconnected from TouchFree Service');
+            }
+        });
+
         this.webSocket.addEventListener('open', this.RequestHandshake, { once: true });
     }